rename `hutchagent/query.py` -> query_builders.py
